Make var-width dict reservations fallible
BaseVarDictDecoder and RleDictVarcharSliceDecoder reserved their
dict-value and aux-entry vectors with infallible Vec::with_capacity.
For a compressed dictionary page the buffer is the decompressed one,
sized by uncompressed_page_size (up to i32::MAX, not bounded by
max_page_size), so num_values can reach ~536M and the reservation
~4x that buffer. On a host that cannot satisfy it, with_capacity
aborts the JVM over JNI -- the same crash class this branch already
fixes elsewhere.

Route both sites, plus the six decimal-dict sites, through one shared
try_reserve_dict_values helper in decoders/mod.rs. It reserves with
try_reserve_exact and maps failure to a recoverable OutOfMemory error,
so an oversized dictionary surfaces a clean error instead of aborting.
The helper's usize::MAX test moves alongside it; the guard in
BaseVarDictDecoder still rejects a tiny-buffer/huge-header page up
front with a precise Layout error.

Also correct the dictionary.rs comment that wrongly attributed the
num_values bound to max_page_size.
